Fixes #99 Adds copy-paste dialog for pull-secret
