+++ crc-tray.csproj
@@ -249,11 +249,11 @@
     <Compile Include="Forms\AboutForm.Designer.cs">
       <DependentUpon>AboutForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Forms\PullSecretPickerForm.cs">
+    <Compile Include="Forms\PullSecretForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Forms\PullSecretPickerForm.Designer.cs">
-      <DependentUpon>PullSecretPickerForm.cs</DependentUpon>
+    <Compile Include="Forms\PullSecretForm.Designer.cs">
+      <DependentUpon>PullSecretForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Helpers\DaemonLauncher.cs" />
     <Compile Include="Helpers\Handlers.cs" />
@@ -287,8 +287,8 @@
     <EmbeddedResource Include="Forms\CrcSettingsForm.resx">
       <DependentUpon>CrcSettingsForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Forms\PullSecretPickerForm.resx">
-      <DependentUpon>PullSecretPickerForm.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\PullSecretForm.resx">
+      <DependentUpon>PullSecretForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\StatusForm.resx">
       <DependentUpon>StatusForm.cs</DependentUpon>
